+++ faiss/include/faiss_C.h
@@ -6,6 +6,7 @@
 #define _FAISS_C_H
 
 #include <stdbool.h>
+#include <stdio.h>
 
 #ifdef __cplusplus
     extern "C"
@@ -22,9 +23,9 @@
         void BuildIndex(FaissProductClusteringDB* fdb, int numOfTrainDataset);
         bool GetTrainStatus(FaissProductClusteringDB* fdb);
         void AddNewVector(FaissProductClusteringDB* fdb, int sizeOfDatabase, float vectors[]);
-        void AddNewVectorWithIDs(FaissProductClusteringDB* fdb, int sizeOfDatabase, float vectors[], long long pids[]);
-        void SearchVector(FaissProductClusteringDB* fdb, int numOfQuery, float vectors[], int kTotal, float distances[], long long pids[]);
-        void SearchVectorByID(FaissProductClusteringDB* fdb, long long pid, float vectors[]);
+        void AddNewVectorWithIDs(FaissProductClusteringDB* fdb, int sizeOfDatabase, float vectors[], int64_t pids[]);
+        void SearchVector(FaissProductClusteringDB* fdb, int numOfQuery, float vectors[], int kTotal, float distances[], int64_t pids[]);
+        void SearchVectorByID(FaissProductClusteringDB* fdb, int64_t pid, float vectors[]);
         void DeleteVectorsByIDs(FaissProductClusteringDB* fdb, int pids[]);
         int GetVectorTotal(FaissProductClusteringDB* fdb);
         void DumpFaissDB(FaissProductClusteringDB* fdb, char fileName[]);
